+++ R/format.R
@@ -103,8 +103,9 @@ time_format_impl <- function(x, format = time_format_default(x), ...) {
         },
         frac = function(x) {
           # Apply time zone offset to x, with truncation for discrete time models.
-          x_tz <- tz_offset(x)
+          chronon <- attr(x, "chronon")
           x <- vec_data(x)
+          x_tz <- tz_offset_impl(x, chronon)
           if(is.integer(x)) x_tz <- trunc(x_tz)
           x <- x + x_tz
           sprintf("%.1f%%", (x - floor(x))*100)